requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.2
django-environ==0.9.0
django-extensions==3.2.3
djangorestframework==3.14.0
fonttools==4.41.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==23.1
pandas==1.5.2
Pillow==9.3.0
protobuf==4.23.3
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
thop==0.1.1.post2209072238
tqdm==4.64.1
typing-extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16